+++ MerapiGolfLogistik/MerapiGolfLogistik/TotalPembelian.cs
@@ -18,9 +18,9 @@ namespace MerapiGolfLogistik
         public string no_nota { get; set; }
         public string nama_barang { get; set; }
         public string nama_kategori { get; set; }
-        public Nullable<int> banyak_barang { get; set; }
-        public Nullable<int> harga_satuan { get; set; }
-        public Nullable<int> total_harga { get; set; }
+        public Nullable<double> banyak_barang { get; set; }
+        public Nullable<double> harga_satuan { get; set; }
+        public Nullable<double> total_harga { get; set; }
         public string satuan { get; set; }
         public System.Guid id_barang { get; set; }
         public Nullable<System.DateTime> tanggal { get; set; }
